chore: clean up propotional farm stuff, getFarmTypeIconByReferredToken, AccordionSponsoredFarm
